YKI(Backend): Move PersonSearchProjection from model to repository
