+++ backend/yki/src/main/java/fi/oph/yki/api/clerk/ClerkCustomerController.java
@@ -47,7 +47,7 @@ public class ClerkCustomerController {
     final int validatedSize = Math.min(size, MAX_PAGE_SIZE);
 
     if (request.personQuery() != null) {
-      final var  personQuery = request.personQuery().trim();
+      final var personQuery = request.personQuery().trim();
       if (personQuery.length() <= 2) {
         throw new IllegalArgumentException("When given the person query, it must have atleast three characters");
       }
