+++ backend/yki/src/main/java/fi/oph/yki/repository/PersonRepository.java
@@ -1,7 +1,6 @@
 package fi.oph.yki.repository;
 
 import fi.oph.yki.model.Person;
-import fi.oph.yki.model.PersonSearchProjection;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
