+++ backend/yki/src/main/java/fi/oph/yki/service/ClerkCustomerService.java
@@ -10,11 +10,11 @@ import fi.oph.yki.api.dto.clerk.ClerkExamDTO;
 import fi.oph.yki.api.dto.clerk.ClerkExamLocationDTO;
 import fi.oph.yki.model.ExamPayment;
 import fi.oph.yki.model.FreeRegistration;
-import fi.oph.yki.model.PersonSearchProjection;
 import fi.oph.yki.model.Registration;
 import fi.oph.yki.onr.OnrService;
 import fi.oph.yki.onr.dto.PersonalDataDTO;
 import fi.oph.yki.repository.PersonRepository;
+import fi.oph.yki.repository.PersonSearchProjection;
 import fi.oph.yki.repository.RegistrationRepository;
 import fi.oph.yki.util.HetuUtils;
 import java.util.Comparator;
